feat(social): 实现用户关注和粉丝功能
- 新增用户关注和粉丝相关的数据结构和接口
- 实现用户关注和取关的功能
- 添加粉丝更新相关的消息和监听器
- 优化关注和粉丝数据的批量处理逻辑
- 新增分布式ID生成类 IDUtil

+++ cshop-social/src/main/java/github/lianyutian/cshop/social/model/po/UserAttention.java
@@ -18,13 +18,13 @@ import lombok.Data;
 public class UserAttention {
   /** 主键id */
   @TableId(type = IdType.AUTO)
-  private Integer id;
+  private Long id;
 
   /** 当前博主用户id */
-  private Integer userId;
+  private Long userId;
 
   /** 关注博主用户id */
-  private Integer attentionId;
+  private Long attentionId;
 
   /** 创建设计，会按这个字段排序 */
   private Date createTime;
